requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bleach==2.1.2
bokeh==0.12.13
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
html5lib==1.0.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-binance==0.5.15
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
zope.interface==4.4.3
